pyproject.toml
[tool.poetry]
name = "flashwise"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2"
djangorestframework = "^3.14.0"
dj-rest-auth = "^3.0.0"
django-allauth = "^0.54.0"
django-cors-headers = "^3.14.0"
coverage = "^7.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
